# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EmotionType = (typeof EmotionType)[keyof typeof EmotionType];

export const EmotionType = {
  stressed: "stressed",
  drunk: "drunk",
  abusive: "abusive",
  pain: "pain",
  calm: "calm",
  unknown: "unknown",
} as const;

export type EmotionAnalysisRequestLanguage =
  (typeof EmotionAnalysisRequestLanguage)[keyof typeof EmotionAnalysisRequestLanguage];

export const EmotionAnalysisRequestLanguage = {
  english: "english",
  tamil: "tamil",
  hindi: "hindi",
} as const;

export interface EmotionAnalysisRequest {
  /** Text transcript to analyze */
  transcript: string;
  language?: EmotionAnalysisRequestLanguage;
  simulatedMode?: boolean;
}

export type EmotionAnalysisResultRiskLevel =
  (typeof EmotionAnalysisResultRiskLevel)[keyof typeof EmotionAnalysisResultRiskLevel];

export const EmotionAnalysisResultRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface EmotionScore {
  emotion: EmotionType;
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
}

export interface EmotionTimelinePoint {
  /** Seconds into the audio */
  timestamp: number;
  emotion: EmotionType;
  /**
   * @minimum 0
   * @maximum 100
   */
  intensity: number;
}

export interface EmotionAnalysisResult {
  primaryEmotion: EmotionType;
  /**
   * @minimum 0
   * @maximum 100
   */
  confidenceScore: number;
  emotionBreakdown: EmotionScore[];
  riskLevel: EmotionAnalysisResultRiskLevel;
  timeline: EmotionTimelinePoint[];
  language: string;
  processingTimeMs: number;
}

export type KeywordDetectionRequestLanguage =
  (typeof KeywordDetectionRequestLanguage)[keyof typeof KeywordDetectionRequestLanguage];

export const KeywordDetectionRequestLanguage = {
  english: "english",
  tamil: "tamil",
  hindi: "hindi",
} as const;

export interface KeywordDetectionRequest {
  transcript: string;
  language?: KeywordDetectionRequestLanguage;
}

export type DetectedKeywordSeverity =
  (typeof DetectedKeywordSeverity)[keyof typeof DetectedKeywordSeverity];

export const DetectedKeywordSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface DetectedKeyword {
  keyword: string;
  position: number;
  severity: DetectedKeywordSeverity;
}

export interface KeywordDetectionResult {
  detectedKeywords: DetectedKeyword[];
  /** Transcript with keyword markers */
  highlightedTranscript: string;
  alertTriggered: boolean;
  alertMessage?: string;
}

export type CallRecordStatus =
  (typeof CallRecordStatus)[keyof typeof CallRecordStatus];

export const CallRecordStatus = {
  active: "active",
  ended: "ended",
  escalated: "escalated",
  resolved: "resolved",
} as const;

export type CallRecordLanguage =
  (typeof CallRecordLanguage)[keyof typeof CallRecordLanguage];

export const CallRecordLanguage = {
  english: "english",
  tamil: "tamil",
  hindi: "hindi",
} as const;

export type CallRecordRiskLevel =
  (typeof CallRecordRiskLevel)[keyof typeof CallRecordRiskLevel];

export const CallRecordRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CallRecord {
  id: number;
  callerId: string;
  callerNumber: string;
  startTime: string;
  endTime?: string;
  /** Duration in seconds */
  duration?: number;
  status: CallRecordStatus;
  language: CallRecordLanguage;
  primaryEmotion: EmotionType;
  riskLevel: CallRecordRiskLevel;
  escalated: boolean;
  transcript?: string;
  detectedKeywords: string[];
  confidenceScore?: number;
  operatorId?: string;
}

export type CreateCallRequestLanguage =
  (typeof CreateCallRequestLanguage)[keyof typeof CreateCallRequestLanguage];

export const CreateCallRequestLanguage = {
  english: "english",
  tamil: "tamil",
  hindi: "hindi",
} as const;

export interface CreateCallRequest {
  callerNumber: string;
  language?: CreateCallRequestLanguage;
  operatorId?: string;
}

export type UpdateCallRequestStatus =
  (typeof UpdateCallRequestStatus)[keyof typeof UpdateCallRequestStatus];

export const UpdateCallRequestStatus = {
  active: "active",
  ended: "ended",
  escalated: "escalated",
  resolved: "resolved",
} as const;

export type UpdateCallRequestRiskLevel =
  (typeof UpdateCallRequestRiskLevel)[keyof typeof UpdateCallRequestRiskLevel];

export const UpdateCallRequestRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdateCallRequest {
  status?: UpdateCallRequestStatus;
  escalated?: boolean;
  primaryEmotion?: EmotionType;
  riskLevel?: UpdateCallRequestRiskLevel;
  transcript?: string;
  detectedKeywords?: string[];
  confidenceScore?: number;
}

export interface EmotionDistributionItem {
  emotion: EmotionType;
  count: number;
  percentage: number;
}

export interface AnalyticsSummary {
  totalCallsToday: number;
  activeCalls: number;
  escalatedCalls: number;
  avgConfidenceScore: number;
  emotionDistribution: EmotionDistributionItem[];
  peakStressHour: string;
  highRiskCallsToday: number;
}

export interface EmotionTrendPoint {
  hour: string;
  stressed: number;
  drunk: number;
  abusive: number;
  pain: number;
  calm: number;
}

export interface CallVolumePoint {
  hour: string;
  count: number;
  highRisk: number;
}