# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionDistributionItem } from "./emotionDistributionItem";

export interface AnalyticsSummary {
  totalCallsToday: number;
  activeCalls: number;
  escalatedCalls: number;
  avgConfidenceScore: number;
  emotionDistribution: EmotionDistributionItem[];
  peakStressHour: string;
  highRiskCallsToday: number;
}