# lib/api-zod/src/generated/types/callRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { CallRecordLanguage } from "./callRecordLanguage";
import type { CallRecordRiskLevel } from "./callRecordRiskLevel";
import type { CallRecordStatus } from "./callRecordStatus";
import type { EmotionType } from "./emotionType";

export interface CallRecord {
  id: number;
  callerId: string;
  callerNumber: string;
  startTime: Date;
  endTime?: Date;
  /** Duration in seconds */
  duration?: number;
  status: CallRecordStatus;
  language: CallRecordLanguage;
  primaryEmotion: EmotionType;
  riskLevel: CallRecordRiskLevel;
  escalated: boolean;
  transcript?: string;
  detectedKeywords: string[];
  confidenceScore?: number;
  operatorId?: string;
}