# lib/api-zod/src/generated/types/callVolumePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface CallVolumePoint {
  hour: string;
  count: number;
  highRisk: number;
}