# lib/api-zod/src/generated/types/createCallRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCallRequestLanguage } from "./createCallRequestLanguage";

export interface CreateCallRequest {
  callerNumber: string;
  language?: CreateCallRequestLanguage;
  operatorId?: string;
}