# lib/api-zod/src/generated/types/detectedKeyword.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { DetectedKeywordSeverity } from "./detectedKeywordSeverity";

export interface DetectedKeyword {
  keyword: string;
  position: number;
  severity: DetectedKeywordSeverity;
}