# lib/api-zod/src/generated/types/emotionAnalysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionAnalysisRequestLanguage } from "./emotionAnalysisRequestLanguage";

export interface EmotionAnalysisRequest {
  /** Text transcript to analyze */
  transcript: string;
  language?: EmotionAnalysisRequestLanguage;
  simulatedMode?: boolean;
}