# lib/api-zod/src/generated/types/emotionScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionType } from "./emotionType";

export interface EmotionScore {
  emotion: EmotionType;
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
}