# lib/api-zod/src/generated/types/emotionTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface EmotionTrendPoint {
  hour: string;
  stressed: number;
  drunk: number;
  abusive: number;
  pain: number;
  calm: number;
}