# lib/api-zod/src/generated/types/emotionType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */

export type EmotionType = (typeof EmotionType)[keyof typeof EmotionType];

export const EmotionType = {
  stressed: "stressed",
  drunk: "drunk",
  abusive: "abusive",
  pain: "pain",
  calm: "calm",
  unknown: "unknown",
} as const;