# lib/api-zod/src/generated/types/updateCallRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emergency Call Emotion Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionType } from "./emotionType";
import type { UpdateCallRequestRiskLevel } from "./updateCallRequestRiskLevel";
import type { UpdateCallRequestStatus } from "./updateCallRequestStatus";

export interface UpdateCallRequest {
  status?: UpdateCallRequestStatus;
  escalated?: boolean;
  primaryEmotion?: EmotionType;
  riskLevel?: UpdateCallRequestRiskLevel;
  transcript?: string;
  detectedKeywords?: string[];
  confidenceScore?: number;
}